backend/requirements.txt
fastapi
uvicorn
biopython
google-generativeai
statsmodels
scipy
numpy
python-dotenv
requests
pydantic